pyproject.toml
[project]
name = "crawl4ai-agent-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.50.0",
    "anyio==4.9.0",
    "argcomplete==3.6.2",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "boto3==1.38.5",
    "botocore==1.38.5",
    "brotli==1.1.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==1.0.7",
    "click==8.1.8",
    "cohere==5.15.0",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "crawl4ai==0.6.2",
    "cryptography==44.0.2",
    "cssselect==1.3.0",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "durationpy==0.9",
    "eval-type-backport==0.2.2",
    "executing==2.2.0",
    "fake-http-header==0.3.5",
    "fake-useragent==2.2.0",
    "fastapi==0.115.9",
    "fastavro==1.10.0",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.39.0",
    "googleapis-common-protos==1.70.0",
    "gotrue==2.12.0",
    "greenlet==3.2.1",
    "griffe==1.7.3",
    "groq==0.23.1",
    "grpcio==1.71.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "humanfriendly==10.0",
    "humanize==4.12.2",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "kubernetes==32.0.1",
    "litellm==1.67.1",
    "logfire==3.14.1",
    "logfire-api==3.14.1",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mcp==1.6.0",
    "mdurl==0.1.2",
    "mistralai==1.7.0",
    "mmh3==5.1.0",
    "more-itertools==10.7.0",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "narwhals==1.37.1",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numpy==2.2.5",
    "oauthlib==3.2.2",
    "onnxruntime==1.21.1",
    "openai==1.76.2",
    "opentelemetry-api==1.32.1",
    "opentelemetry-exporter-otlp-proto-common==1.32.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.32.1",
    "opentelemetry-exporter-otlp-proto-http==1.32.1",
    "opentelemetry-instrumentation==0.53b1",
    "opentelemetry-instrumentation-asgi==0.53b1",
    "opentelemetry-instrumentation-fastapi==0.53b1",
    "opentelemetry-proto==1.32.1",
    "opentelemetry-sdk==1.32.1",
    "opentelemetry-semantic-conventions==0.53b1",
    "opentelemetry-util-http==0.53b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==10.4.0",
    "playwright==1.52.0",
    "pluggy==1.5.0",
    "postgrest==1.0.1",
    "posthog==4.0.1",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-ai==0.1.8",
    "pydantic-ai-slim==0.1.8",
    "pydantic-core==2.33.2",
    "pydantic-evals==0.1.8",
    "pydantic-graph==0.1.8",
    "pydantic-settings==2.9.1",
    "pydeck==0.9.1",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyopenssl==25.0.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "pytest==8.3.5",
    "pytest-mock==3.14.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "qdrant-client[fastembed]>=1.14.1",
    "rank-bm25==0.2.2",
    "realtime==2.4.3",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==14.0.0",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "s3transfer==0.12.0",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==4.1.0",
    "setuptools==80.0.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "snowballstemmer==2.2.0",
    "soupsieve==2.7",
    "sse-starlette==2.3.3",
    "starlette==0.45.3",
    "storage3==0.11.3",
    "streamlit==1.45.0",
    "strenum==0.4.15",
    "supabase==2.15.1",
    "supafunc==0.9.4",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tf-playwright-stealth==1.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "torch==2.7.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "typer==0.15.3",
    "types-requests==2.32.0.20250328",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "watchdog==6.0.0",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "wrapt==1.17.2",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zipp==3.21.0",
]
